fix/format: fix parallel omp statement and format docs architecture
